+++ src/me/sirfaizdat/prison/mines/cmds/CmdInfo.java
@@ -24,7 +24,7 @@ import me.sirfaizdat.prison.core.MessageUtil;
 import me.sirfaizdat.prison.core.Prison;
 import me.sirfaizdat.prison.mines.Mines;
 import me.sirfaizdat.prison.mines.Block;
-import me.sirfaizdat.prison.mines.Mine;
+import me.sirfaizdat.prison.mines.entities.Mine;
 
 import java.util.Map;
