+++ src/me/sirfaizdat/prison/mines/cmds/CmdRedefine.java
@@ -23,7 +23,7 @@ import me.sirfaizdat.prison.core.Command;
 import me.sirfaizdat.prison.core.MessageUtil;
 import me.sirfaizdat.prison.core.Prison;
 import me.sirfaizdat.prison.mines.Mines;
-import me.sirfaizdat.prison.mines.Mine;
+import me.sirfaizdat.prison.mines.entities.Mine;
 
 /**
  * @author SirFaizdat
