+++ src/me/sirfaizdat/prison/mines/cmds/CmdReset.java
@@ -21,7 +21,7 @@ package me.sirfaizdat.prison.mines.cmds;
 import me.sirfaizdat.prison.core.Command;
 import me.sirfaizdat.prison.core.MessageUtil;
 import me.sirfaizdat.prison.mines.Mines;
-import me.sirfaizdat.prison.mines.Mine;
+import me.sirfaizdat.prison.mines.entities.Mine;
 
 /**
  * @author SirFaizdat
