+++ src/me/sirfaizdat/prison/mines/events/MineResetEvent.java
@@ -19,7 +19,7 @@
 
 package me.sirfaizdat.prison.mines.events;
 
-import me.sirfaizdat.prison.mines.Mine;
+import me.sirfaizdat.prison.mines.entities.Mine;
 import org.bukkit.event.Event;
 import org.bukkit.event.HandlerList;
